feat: Implement hooks for fetching NFL and CFB game data, player stats, and team statistics
- Added `useLastTeamGame` hook to fetch the most recent completed game for NFL and CFB teams.
- Introduced `useNFLGameDetails` hook to retrieve detailed information about NFL games, including officials and scoring plays.
- Created `useNFLPlayerCareerStats` hook for fetching player career statistics across multiple seasons.
- Developed `useGameOfficials` hook to fetch game officials and injury reports for NBA and College Basketball.
- Implemented `useHomeData` hook to manage home screen data, including games and news.
- Added `useTeamStats` hook for fetching aggregated team statistics for the NBA.
- Created styles for player headers in `PlayerHeader.styles.ts`.
- Defined new types for CBB players in `cbb.ts`.

+++ components/NFL/Games/NFLGameSquareCard.tsx
@@ -144,7 +144,7 @@ function NFLGameSquareCard({ game, isDark }: Props) {
 
   // --- Date formatting ---
   const formattedDate = gameDate
-    ? gameDate.toLocaleDateString("en-US", { month: "numeric", day: "numeric" })
+    ? gameDate.toLocaleDateString("en-US", { month: "short", day: "numeric" })
     : "";
 
   const formattedTime = gameDate
